Adding Killer to index.html (#146)
Adding at the request of a user in the Mojira Discord server - it appears to be the cause of the "failed to download file" errors that have plagued users for the last year. A workaround for the issue has existed in MCL-23025.
I discovered this yesterday and was succeeded in assisting someone fix their failed to download file error with this.
The credit for finding this should go to u/NoOneCanCode on Reddit.